feat(url shortening): adds persistence of shortened urls
By adding persistence to the shortened urls, we're now able to check if a requested url shortened
has already been done. If it has, then we return it and doesn't generate another entry on the
database. If it hasn't, then we proceed with the normal flow.

#1

+++ csharp/src/BeTiny.Api/Domain/ValueObjects/Counter.cs
@@ -6,7 +6,7 @@ namespace BeTiny.Api.Domain.ValueObjects
 {
     public class Counter : ValueObject
     {
-        public int Value { get; private set; }
+        public long Value { get; private set; }
 
         public Counter()
         {
@@ -14,7 +14,7 @@ namespace BeTiny.Api.Domain.ValueObjects
         }
 
         [JsonConstructor]
-        public Counter(int value)
+        public Counter(long value)
         {
             Value = value;
         }
